feat: added 2 custom exception, make exception handling to Invoice endpoints task: exception handling at Transaction.

+++ src/main/java/com/zk/budgetmate/service/InvoiceServiceImp.java
@@ -1,6 +1,8 @@
 package com.zk.budgetmate.service;
 
 import com.zk.budgetmate.DTO.InvoiceDTO;
+import com.zk.budgetmate.exception.DuplicateResourceException;
+import com.zk.budgetmate.exception.ResourceNotFoundException;
 import com.zk.budgetmate.mapper.InvoiceMapper;
 import com.zk.budgetmate.model.Invoice;
 import com.zk.budgetmate.repository.InvoiceRepository;
@@ -8,7 +10,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -31,17 +32,24 @@ public class InvoiceServiceImp implements InvoiceService {
   }
 
   @Override
-  public Optional<InvoiceDTO> getInvoiceById(Long id) {
+  public InvoiceDTO getInvoiceById(Long id) {
     return invoiceRepository.findById(id)
-        .map(invoiceMapper::toDTO);
+        .map(invoiceMapper::toDTO)
+        .orElseThrow(() -> new ResourceNotFoundException("Invoice not found with the given id: " + id));
   }
 
   @Override
   public InvoiceDTO saveNewInvoice(InvoiceDTO dto) {
+    if (invoiceRepository.existsByName(dto.getName())) {
+      throw new DuplicateResourceException("Invoice with the given name: " + dto.getName() + " is already exists");
+    }
     return invoiceMapper.toDTO(invoiceRepository.save(invoiceMapper.toEntity(dto)));
   }
 
   public void deleteInvoiceById(Long id) {
+    if (!invoiceRepository.existsById(id)) {
+      throw new ResourceNotFoundException("Invoice not found with the given id: " + id);
+    }
     invoiceRepository.deleteById(id);
   }
